+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-# Webrpc changelog
+# webrpc v0.9.0
 
-## v0.9.0
+Towards reaching webrpc@v1.0.0, we have decided to make some changes to webrpc schema and RIDL file format.
 
-Towards reaching webrpc@v1.0.0, we have decided to make some minor changes to RIDL file format. Here's what you need to do to migrate to webrpc@v0.9.0+.
+## Breaking changes
 
-## RIDL v0.9.0 changes
+### RIDL v0.9.0 changes
 
-Keyword "message" was replaced by "struct".
+Keyword "message" was renamed to "struct".
 
 ```diff
  webrpc = v1
@@ -21,9 +21,21 @@ Keyword "message" was replaced by "struct".
    - createdAt?: timestamp
 ```
 
-## RIDL v0.9.0 migration guide
+## JSON schema v0.9.0 changes
 
-You can run this command to migrate your RIDL files to webrpc@v0.9.0+:
+Field "messages" was renamed to "types".
+
+```diff
+{
+
+}
+```
+
+## Migration guide
+
+### RIDL v0.9.0 migration guide
+
+Run this command to migrate your RIDL files to webrpc@v0.9.0+:
 
 ```bash
 #!/bin/bash
@@ -31,12 +43,23 @@ You can run this command to migrate your RIDL files to webrpc@v0.9.0+:
 find . -name '*.ridl' -exec sed -i -e 's/^message /struct /g' {} \;
 ```
 
-## JSON schema v0.9.0 migration guide
+### JSON schema v0.9.0 migration guide
 
-You can run this command to migrate your .json schema files to webrpc@v0.9.0+:
+Run this command to migrate your `.json` schema files to webrpc@v0.9.0+:
 
 ```bash
 #!/bin/bash
 
-find . -name '*.json' -exec sed -i -e 's/^"messages":/\"structs\":/g' {} \;
+find . -name '*.json' -exec sed -i -e 's/"messages":/\"types\":/g' {} \;
 ```
+
+### Generator templates v0.9.0 migration guide
+
+Run this command to migrate your `.go.tmpl` templates to webrpc@v0.9.0+:
+
+```bash
+#!/bin/bash
+
+find . -name '*.go.tmpl' -exec sed -i -e 's/.Messages:/\".Types\":/g' {} \;
+```
+